refactor: split git watcher tests
